+++ CHANGELOG.md
@@ -645,3 +645,5 @@
 * Define `_j` operator
 * Define `j_` operator
 * Define `yoo` operator
+* Define `yi_cn_dp` operator
+* Define `yi_lm_dp` operator